chore: file to updated and add contact renamed
